feat #36: agrego la selección a sede
Elmino pedido.py yañado a sede la logica de negocio
